Use *ngIf to mange access to the selectedHero
When the app starts the selectedHero is undefined which causes an error and prevents the app from loading. This issue also exists when the data is sometimes not present due to the async nature of NodeJS. Using a conditional directive solves the problem by only displaying the output when it exists.