Setup the base service to handle hero data
To decouple the data access from the components that use it requires a service module. This service will be responsible for accessing the data and providing it upon request. This way the components  that use the data are not concerned with how the data is accessed  or where the data originates.